feat:added fetch post and other related helper function to track like,comment

+++ controller/likecommentsharecontroller.go
@@ -3,9 +3,12 @@ package controller
 import (
 	"cometosee/common"
 	"cometosee/firebase"
+	"cometosee/model"
 	"fmt"
 	"net/http"
 	"time"
+
+	"cloud.google.com/go/firestore"
 )
 
 func LikePost(w http.ResponseWriter, r *http.Request) {
@@ -31,9 +34,9 @@ func LikePost(w http.ResponseWriter, r *http.Request) {
 
 	likeRef := client.Collection("posts").Doc(body.PostId).Collection("likes").Doc(body.UserId)
 
-	doc, _ := likeRef.Get(ctx)
+	doc, err := likeRef.Get(ctx)
 
-	if doc.Exists() {
+	if err == nil && doc.Exists() {
 		likeRef.Delete(ctx)
 		fmt.Fprint(w, "unliked")
 		return
@@ -44,6 +47,25 @@ func LikePost(w http.ResponseWriter, r *http.Request) {
 		"created": time.Now(),
 	})
 	fmt.Fprint(w, "liked")
+}
+
+// counting the like helper funcc
+func LikeCount(postId string) int {
+	client := firebase.Firestore()
+	defer client.Close()
+
+	likescount := client.Collection("posts").Doc(postId).Collection("likes").Documents(ctx)
+
+	count := 0
+
+	for {
+		_, err := likescount.Next()
+		if err != nil {
+			break
+		}
+		count++
+	}
+	return count
 }
 
 func CommentPost(w http.ResponseWriter, r *http.Request) {
@@ -81,6 +103,51 @@ func CommentPost(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, "comment added")
 }
 
+// helper func for  countinngg cmt
+func CommentCount(postId string) int {
+	client := firebase.Firestore()
+	defer client.Close()
+
+	commentsCount := client.Collection("posts").Doc(postId).Collection("comments").Documents(ctx)
+
+	count := 0
+	for {
+		_, err := commentsCount.Next()
+		if err != nil {
+			break
+		}
+		count++
+	}
+	return count
+}
+
+// fetch comment helper funcccc
+func FecthComment(postId string) []model.Comment {
+	client := firebase.Firestore()
+	defer client.Close()
+
+	fetchcmt := client.Collection("posts").Doc(postId).Collection("comments").OrderBy("created", firestore.Asc).Documents(ctx)
+
+	var comments []model.Comment
+
+	for {
+		doc, err := fetchcmt.Next()
+		if err != nil {
+			break
+		}
+		data := doc.Data()
+		comments = append(comments, model.Comment{
+			ID:      doc.Ref.ID,
+			UserID:  data["user_id"].(string),
+			Comment: data["comment"].(string),
+			Created: data["created"],
+		})
+	}
+	return comments
+
+}
+
+// barun donottt forget to make shareable link fromfrontend
 func SharePost(w http.ResponseWriter, r *http.Request) {
 	type Reqbody struct {
 		PostId string `json:"post_id"`
@@ -110,10 +177,10 @@ func SharePost(w http.ResponseWriter, r *http.Request) {
 	})
 
 	if err != nil {
-		common.WriteJSONError(w, "failed to save comment", http.StatusInternalServerError)
+		common.WriteJSONError(w, "failed to save share", http.StatusInternalServerError)
 		return
 	}
 
-	fmt.Fprintf(w, "comment added")
+	fmt.Fprintf(w, "shared added")
 
 }
